oneplus3: rework dalvik heap configuration
The previous configuration (from Stock Oxygen OS) doesn't pass the check:

heapgrowthlimit: 256m
heapsize: 512m
heaptargetutilization: 0.75

0.75/2 = 0.375

To pass the check this has to be true:

0.375*512 = 256 (WRONG)

Check not passed.

The new configuration I made is more optimized for higher RAM devices and passes the check:

heapgrowthlimit: 384m
heapsize: 1024m
heaptargetutilization: 0.75

0.75/2 = 0.375

To pass the check this has to be true:

0.375*1024 = 384 (TRUE)

Check passed.

This patch implements the new configuration.

Change-Id: I6839339382229da80546761c3746a032081ff2cd
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>